feat(userproducts): replace products controller with user_products controller
