+++ app/views/user_products/_row.html.erb
