+++ app/views/user_products/edit.html.erb
