+++ app/views/user_products/index.html.erb
