+++ spec/controllers/user_products_controller_spec.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+RSpec.describe UserProductsController, type: :controller do
+
+end
